Remove QueryError exception because it's never raised to the user
When a QueryError exceptions happens, it's because retries are enabled on the AutoExtract client.
If a QueryError exception is raised, Tenacity is capturing it and retrying it.
If it's not possible to retry a request due to a QueryError, Tenacity raised a QueryRetryError exception.

In short, the QueryError never reaches userland's code.
QueryRetryError does.